# doc/openapi-reference.yml
openapi: "3.0.3"
info:
  title: "Web Application Scanner"
  description: "
    A wrapper for the Zed Attack Proxy (ZAP) web application scanner

    This version of the API specification is a static reference defining
    the required endpoints and data fields.

    For an up-to-date documentation you can check the `/doc/openapi.yml`
    endpoint served by the main service.
    "
  contact:
    name: "Greenbone AG"
    url: "https://www.greenbone.net/"
  license:
    name: "AGPL-3.0-or-later"
    url: "https://spdx.org/licenses/AGPL-3.0-or-later.html"
  version: "0.1"
servers:
  - url: /
...
paths:
...
  /scans:
    head:
      description: "Get the response header. It contains the API version and available authentication methods."
      operationId: "get_info_auth"
      tags:
        - "general"
      responses:
        "204":
          headers:
            api-version:
              description: "Comma separated list of available API versions"
              schema:
                type: "string"
            authentication:
              description: "Supported authentication methods"
              schema:
                type: "string"
          description: "Authenticated and authorized"
        "401":
          headers:
            api-version:
              description: "Comma separated list of available API versions"
              schema:
                type: "string"
            authentication:
              description: "Supported authentication methods"
              schema:
                type: "string"
          description: "Unauthorized. Required or invalid client certificates"
    post:
      description: "This request just creates the scan. It can be started afterwards with the scan_action request."
      operationId: "create_scan"
      tags:
        - "scan"
      requestBody:
        description: "Scan to add"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanReq"
            examples:
              schema:
                description: "Schema of a Scan."
              create simple scan:
                $ref: "#/components/examples/scan_simple"
              create complex scan:
                $ref: "#/components/examples/scan_full_req"
      responses:
        "201":
          description: "Scan created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanID"
              examples:
                created scan:
                  $ref: "#/components/examples/scan_id"
        "400":
          description: "Bad Request body"
        "403":
          description: "ScanID already in use"
...
  /scans/{id}:
    get:
      description: "Get a scan from the scan manager."
      operationId: "get_scan"
      tags:
        - "scan"
      parameters:
        - $ref: "#/components/parameters/ScanID"
      responses:
        "200":
          description: "Get Scan"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResp"
              examples:
                scan:
                  $ref: "#/components/examples/scan_full_resp"

        "404":
          description: "Scan not found"
    post:
      description: "Perform an action to a scan."
      operationId: "scan_action"
      tags:
        - "scan"
      parameters:
        - $ref: "#/components/parameters/ScanID"
      requestBody:
        description: "Action to perform."
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanAction"
            examples:
              start scan:
                $ref: "#/components/examples/scan_action_start"
              stop scan:
                $ref: "#/components/examples/scan_action_stop"
      responses:
        "204":
          description: "Action performed"
        "400":
          description: "Bad request body"
        "404":
          description: "Scan not found"
        "406":
          description: "Unable to perform action because of the current scan status"
        "501":
          description: "Action not supported"
    delete:
      description: "Delete a scan from the scan manager."
      operationId: "delete_scan"
      tags:
        - "scan"
      parameters:
        - $ref: "#/components/parameters/ScanID"
      responses:
        "204":
          description: "Scan deleted"
        "404":
          description: "Scan not found"
        "406":
          description: "A running scan cannot be deleted"
  /scans/{id}/results:
    get:
      description: "Get results from a scan."
      operationId: "get_results"
      tags:
        - "scan"
      parameters:
        - $ref: "#/components/parameters/ScanID"
        - name: range
          in: query
          description: "Get a range of results (e.g. `0-12`).
            In case only a single number is given (e.g. `13`), all available results from this index on are returned.
            If the scan is still running, new results will occur and must be collected with new request starting from the last processed index.
            If no results are in the given range, an empty array is returned."
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "A list of results"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Result"
              examples:
                schema:
                  description: "Schema of a list of results response"
                get results 0-3:
                  $ref: "#/components/examples/scan_results"
        "400":
          description: "Bad range format"
        "404":
          description: "Scan not found"
        "406":
          description: "A scan, that has not started, does not contain results"
  /scans/{id}/results/{rid}:
    get:
      description: "Get a specific result from the scan."
      operationId: "get_result"
      tags:
        - "scan"
      parameters:
        - $ref: "#/components/parameters/ScanID"
        - $ref: "#/components/parameters/ResultID"
      responses:
        "200":
          description: "The requested result"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
              examples:
                schema:
                  description: "Schema of a result"
                get result 3:
                  $ref: "#/components/examples/scan_result"
                host detail:
                  $ref: "#/components/examples/host_detail"
        "404":
          description: "Result or Scan not found"
        "406":
          description: "A scan that has not started, cannot contain results"
  /scans/{id}/status:
    get:
      description: "Get the current status of a scan."
      operationId: "get_scan_status"
      tags:
        - "scan"
      parameters:
        - $ref: "#/components/parameters/ScanID"
      responses:
        "200":
          description: "The requested status"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
              examples:
                schema:
                  description: "Schema of a status response."
                status of a stored scan:
                  $ref: "#/components/examples/scan_status_stored"
                status of a running scan:
                  $ref: "#/components/examples/scan_status_running"
                status of a succeeded scan:
                  $ref: "#/components/examples/scan_status_success"
                status of a failed scan:
                  $ref: "#/components/examples/scan_status_fail"
        "404":
          description: "Scan not found"
components:
  parameters:
    ScanID:
      name: id
      in: path
      description: "ID of a Scan"
      required: true
      schema:
        type: "string"
    ResultID:
      name: rid
      in: path
      description: "ID of a Result"
      required: true
      schema:
        type: "string"
  schemas:
    ScanID:
      description: "A scan ID to identify a scan."
      type: "string"
    ScanReq:
      description: "Model representing a scan request."
      type: "object"
      properties:
        scan_id:
          description: "An optional scan ID to identify the scan. If not given, a random UUID will be generated and returned in the response."
          type: "string"
        target:
          $ref: "#/components/schemas/Target"
        scan_preferences:
          description: "Overwrite the default settings of the Scanner."
          type: "array"
          items:
            $ref: "#/components/schemas/ScannerPreference"
        vts:
          type: "array"
          description: "A collection of VTs, which are run for the given target."
          items:
            $ref: "#/components/schemas/VT"
      required:
        - target
        - vts
    ScanResp:
      description: "Model representing a scan response."
      type: "object"
      properties:
        scan_id:
          $ref: "#/components/schemas/ScanID"
        target:
          $ref: "#/components/schemas/Target"
        scan_preferences:
          description: "Overwrite the default settings of the Scanner."
          type: "array"
          items:
            $ref: "#/components/schemas/ScannerPreference"
        vts:
          type: "array"
          description: "A collection of VTs, which are run for the given target."
          items:
            $ref: "#/components/schemas/VT"
      required:
        - target
        - vts
...
    Result:
      description: "A result of a Scan"
      type: "object"
      properties:
        id:
          description: "An ID, which is unique for the scan. It is auto incremental and starts at 0 for the first result."
          type: "integer"
          format: "int32"
        type:
          description: "The type of the result."
          type: "string"
          enum:
            - alarm
            - log
            - error
            - host_start
            - host_stop
            - host_detail
        ip_address:
          description: "The IPv4 or IPv6 of the target the result was found."
          type: "string"
        hostname:
          description: "The hostname of the target the result was found."
          type: "string"
        oid:
          description: "The identifier of the VT in which found the result."
          type: "string"
        port:
          description: "The port that was used to find the result."
          type: "integer"
          format: "int32"
        protocol:
          description: "The protocol that was used to find the result, corresponding to the port."
          type: "string"
          enum:
            - udp
            - tcp
        message:
          description: "Additional information about the result."
          type: "string"
        detail:
          description: "The detail object is only used for results of type host_detail. It contains information about a scanned hosted such as hardware information, architecture and many more."
          type: "object"
          properties:
            name:
              description: "descriptive name of the detail"
              type: "string"
            value:
              description: "value of the host detail"
              type: "string"
            source:
              description: "source of the host detail"
              type: "object"
              properties:
                type:
                  description: "type of the source, e.g. nvt"
                  type: "string"
                name:
                  description: "descriptive name of the source, in case of type nvt this is its OID"
                  type: "string"
                description:
                  description: "Optional information about the detection, could contain how the data actually was collected."
                  type: "string"
              required:
                - type
                - name
          required:
            - name
            - value
            - source

      required:
        - type
...
    ScanAction:
      description: "An action to perform on a scan"
      type: "object"
      properties:
        action:
          description: "The action to perform"
          type: "string"
          enum:
            - start
            - stop
      required:
        - "action"
...
  examples:
    scan_simple:
      description: "A simple example for creating a scan."
      value:
        {
          "target":
            {
              "hosts": ["127.0.0.1"],
              "ports": [{ "range": [{ "start": 22 }] }],
            },
          "vts": [{ "oid": "1.3.6.1.4.1.25623.1.0.10267" }],
        }
    scan_full_req:
      description: "A complex example for creating a scan, that uses all available fields."
      value:
        {
          "target":
            {
              "hosts":
                [
                  "127.0.0.1",
                  "192.168.0.1-15",
                  "10.0.5.0/24",
                  "::1",
                  "2001:db8:0000:0000:0000:0000:0000:0001-00ff",
                  "2002::1234:abcd:ffff:c0a8:101/64",
                  "examplehost",
                ],
              "excluded_hosts": ["192.168.0.14"],
              "ports":
                [
                  {
                    "protocol": "udp",
                    "range": [{ "start": 22 }, { "start": 1024, "end": 1030 }],
                  },
                  { "protocol": "tcp", "range": [{ "start": 24, "end": 30 }] },
                  { "range": [{ "start": 100, "end": 1000 }] },
                ],
              "credentials":
                [
                  {
                    "service": "ssh",
                    "port": 22,
                    "usk":
                      {
                        "username": "user",
                        "password": "pw",
                        "private": "ssh-key...",
                      },
                  },
                  {
                    "service": "smb",
                    "up": { "username": "user", "password": "pw" },
                  },
                  {
                    "service": "snmp",
                    "snmp":
                      {
                        "username": "user",
                        "password": "pw",
                        "community": "my_community",
                        "auth_algorithm": "md5",
                        "privacy_password": "priv_pw",
                        "privacy_algorithm": "aes",
                      },
                  },
                ],
              "alive_test_ports":
                [
                  { "protocol": "tcp", "range": [{ "start": 1, "end": 100 }] },
                  { "range": [{ "start": 443 }] },
                ],
              "alive_test_methods":
                ["icmp", "tcp_syn", "tcp_ack", "arp", "consider_alive"],
              "reverse_lookup_unify": true,
              "reverse_lookup_only": false,
            },
          "scan_preferences":
            [
              { "id": "target_port", "value": "443" },
              { "id": "use_https", "value": "1" },
              { "id": "profile", "value": "fast_scan" },
            ],
          "vts":
            [
              {
                "oid": "1.3.6.1.4.1.25623.1.0.10662",
                "parameters":
                  [{ "id": 1, "value": "200" }, { "id": 2, "value": "yes" }],
              },
              { "oid": "1.3.6.1.4.1.25623.1.0.10330" },
            ],
        }
    scan_full_resp:
      description: "A complex example for creating a scan, that uses all available fields."
      value:
        {
          "scan_id": "6c591f83-8f7b-452a-8c78-ba35779e682f",
          "target":
            {
              "hosts":
                [
                  "127.0.0.1",
                  "192.168.0.1-15",
                  "10.0.5.0/24",
                  "::1",
                  "2001:db8:0000:0000:0000:0000:0000:0001-00ff",
                  "2002::1234:abcd:ffff:c0a8:101/64",
                  "examplehost",
                ],
              "ports":
                [
                  {
                    "protocol": "udp",
                    "range": [{ "start": 22 }, { "start": 1024, "end": 1030 }],
                  },
                  { "protocol": "tcp", "range": [{ "start": 24, "end": 30 }] },
                  { "range": [{ "start": 100, "end": 1000 }] },
                ],
              "credentials":
                [
                  {
                    "service": "ssh",
                    "port": 22,
                    "usk":
                      {
                        "username": "user",
                        "password": "pw",
                        "private": "ssh-key...",
                      },
                  },
                  {
                    "service": "smb",
                    "up": { "username": "user", "password": "pw" },
                  },
                  {
                    "service": "snmp",
                    "snmp":
                      {
                        "username": "user",
                        "password": "pw",
                        "community": "my_community",
                        "auth_algorithm": "md5",
                        "privacy_password": "priv_pw",
                        "privacy_algorithm": "aes",
                      },
                  },
                ],
              "alive_test_ports":
                [
                  { "protocol": "tcp", "range": [{ "start": 1, "end": 100 }] },
                  { "range": [{ "start": 443 }] },
                ],
              "alive_test_methods":
                ["icmp", "tcp_syn", "tcp_ack", "arp", "consider_alive"],
              "reverse_lookup_unify": true,
              "reverse_lookup_only": false,
            },
          "scan_preferences":
            [
              { "id": "target_port", "value": "443" },
              { "id": "use_https", "value": "1" },
              { "id": "profile", "value": "fast_scan" },
            ],
          "vts":
            [
              {
                "oid": "1.3.6.1.4.1.25623.1.0.10662",
                "parameters":
                  [{ "id": 1, "value": "200" }, { "id": 2, "value": "yes" }],
              },
              { "oid": "1.3.6.1.4.1.25623.1.0.10330" },
            ],
        }
    scan_id:
      description: "The ID of the created scan"
      value: "6c591f83-8f7b-452a-8c78-ba35779e682f"
    scan_action_start:
      description: "Start a scan"
      value: { "action": "start" }
    scan_action_stop:
      description: "Stop a running scan"
      value: { "action": "stop" }
    scan_results:
      description: "Example for getting results.
        The query range for this example could be either

        - 0-3

        - 0

        - missing


        In case it is 0 or missing the running scan has only generated 4 results so far."
      value:
        [
          {
            "id": 0,
            "type": "host_start",
            "ip_address": "127.0.0.1",
            "port": 22,
            "protocol": "tcp",
            "message": "Thu Mar 23 15:16:37 2023",
          },
          {
            "id": 1,
            "type": "error",
            "ip_address": "127.0.0.1",
            "hostname": "localhost",
            "protocol": "tcp",
            "message": "MQTT initialization failed",
          },
          {
            "id": 2,
            "type": "log",
            "ip_address": "127.0.0.1",
            "hostname": "localhost",
            "oid": "1.3.6.1.4.1.25623.1.0.117628",
            "port": 22,
            "protocol": "tcp",
            "message": "FTP is enabled on the remote SSH service.",
          },
          {
            "id": 3,
            "type": "alarm",
            "ip_address": "127.0.0.1",
            "hostname": "localhost",
            "oid": "1.3.6.1.4.1.25623.1.0.147696",
            "protocol": "tcp",
            "message": "Installed version: 9.53.3\nFixed version:     9.55\nInstallation\npath / port:       /usr/bin/gs",
          },
        ]
    scan_result:
      description: "Example for getting a single result. The Result ID in the request path was set to 3."
      value:
        {
          "id": 3,
          "type": "alarm",
          "ip_address": "127.0.0.1",
          "hostname": "localhost",
          "oid": "1.3.6.1.4.1.25623.1.0.147696",
          "port": 22,
          "protocol": "tcp",
          "message": "Installed version: 9.53.3\nFixed version:     9.55\nInstallation\npath / port:       /usr/bin/gs",
        }
    host_detail:
      description: "Example for a result of type host detail"
      value:
        {
          "id": 3,
          "type": "host_detail",
          "ip_address": "127.0.0.1",
          "hostname": "localhost",
          "oid": "1.3.6.1.4.1.25623.1.0.103997",
          "detail":
            {
              "name": "OS",
              "value": "Debian GNU/Linux 11.5",
              "source":
                {
                  "type": "nvt",
                  "name": "1.3.6.1.4.1.25623.1.0.50282",
                  "description": "Determine OS and list of installed packages via SSH login",
                },
            },
        }
    scan_status_stored:
      description: "Status of a queued Scan"
      value: { "status": "stored" }
    scan_status_running:
      description: "Status of a running Scan"
      value:
        {
          "start_time": 1679649183,
          "status": "running",
          "host_info":
            {
              "all": 14,
              "excluded": 0,
              "dead": 4,
              "alive": 6,
              "queued": 1,
              "finished": 1,
              "scanning": [
                "127.0.0.1": 80,
                "10.0.5.1": 55,
                "10.0.5.3": 40
              ]
            },
        }
...